fix(gui): fix a miss-formatting of registers edit-box values
In hex mode, when one leaves the edit box w/o any changes, a crapped value with an exceeding '0x' prefix
is inserted. We restore now the last input value. Besides, in float mode, a miss-formated crapped value
with too many zero is inserted.
For all modes, we simply now restore the last input value instead of replacing the value by zero

style(gui): attempt to enhance the layout for the float mode registers values.

Change the registers edit-box values format as too many values cannot be decently filled-in.
Remove the unconditionnal 5 char limit. We use now a 'G' or 'G5' format (if the value length exceeds the
editbox size)