refactor(ui): refactor and strengthen IContainer 'components' attribute, among BaseForm and child forms
WinForms inheritance isn't robust and leads to hassles between the design mode and runtime mode. Manual
explicit code in "reserved" designer region may or may not work, may or may not be lost.
The pollTimer member is removed from the designer scope, and moved to the MasterForm code area.
A Tooltip member is added at BaseForm level (designer realm).