feature(ui): DataTab - Several enhancements to the textboxes user input UX
Add validation to Address and Size fields.
Refactor the user inputs data flow (add a 'dirty' flag, reduce the ui updates reentrance, prevent from
uncontrolled data input paths)
Optimize the textboxes layout
StartAddress supports hexadecimal/decimal input
The register editboxes that overtake the scope/extent of the actual modbus (multiple)operation are dimmed
and made read-only like.
The current edited register input is rendered with a highlight state

BREAKING CHANGES: integer display render signed values (thus unsigned may be inputted too)

feature(ui/data): Attach an object context per textbox and register value

Adding new entities: MBDataItem{T} and derived classes. The 'data-item' cache the value and the formatting
text value related to. It's attached to the textBox. The value formatting and parsing process is moved
from the UI forms and spared methods to those new entities (thus reducing the code friction)

feature(ui) Some changes onto the LedBulb

The user control now copes with the enabled/disabled state of the control
The LedBulb class is updated to render a disabled/dimmed view of the graphics-path of the 'led' shape,
according to the enabled-state of the control.

BREAKING CHANGES: (or fix) the bit set/unset didn't seem to be accurate. Led/bits are now aligned with the
word boundary (i.e there is a strict bijection with the binary display mode). Adressing is modified. A
led/bit address is referenced as the word address, concatenated with the bit index (register-offset.bit-i)

feature(settings): Add an application option for the default addressing display/input mode

The global option AddrIsHexaFormatted is added to application settings, and is adjustable at command line
with parameter --hexAddr {[true]|false}
This option is used for address global offset and address label of each mbus register

feature(ui): Behavior clarification and enhancement about floating point encoding

The BE/LE encoding pane is disabled when display-mode isn't set to float32.
A checkbox 'swap buffer' is added to the BE/LE encoding pane. It enables to choose if the buffer needs to be
swapped when an endianness change occurs (only the registers active subset is involved). A warning and
message box explain the process when change is engaged by the user.
Besides, another checkbox is added to swap the data table (word) endianness (see the float32 display mode
area). This 2nd word-swapping only involved the text boxes, indeed it does not affect the underlying data
but only the displaying of float values, and it's supplied as a convenience especially for the scanner at
reading side (while exploring the received data)

fix(tab/ui): The current tab may be displayed with a wrong format (discrepancy between the Tab and the main form)

When a tab is selected, the globally selected display/format is updated to reflect the current format of the
settled registers within this tab (which allow to actually use a different format per tab)

fix(tab/ui): Tab data table isn't updated when the content related to has been modified from another tab

When several tabs overlap or refer to the same adress range, the data table content could not be accurate,
if the table format wasn't changed.
The data table is now updated at each tab activation

fix(tab/ui): The tab caption might not be up to date after an address/size change (need to switch tabs)

The button Apply event handler management was hazardous and incomprehensible